fix auth

+++ controllers/blog.controller.js
@@ -100,7 +100,7 @@ const blogController = {
           message: "Blog not found by this Id",
         });
 
-      if (blog.author.toString() !== req.user.id && req.user.role !== "admin") {
+      if (req.user.role !== "admin") {
         return res.status(403).json({
           message: "You are not authorized to update this blog",
         });
@@ -127,14 +127,11 @@ const blogController = {
         return res.status(404).json({
           message: "Blog not found by this Id",
         });
-         if (
-           blog.author.toString() !== req.user.id &&
-           req.user.role !== "admin"
-         ) {
-           return res.status(403).json({
-             message: "You are not authorized to update this blog",
-           });
-         }
+      if (req.user.role !== "admin") {
+        return res.status(403).json({
+          message: "You are not authorized to update this blog",
+        });
+      }
       await Blog.findByIdAndDelete(req.params.blogId);
       res.status(200).json({
         status: "success",
